feat: create info block
