+++ src/components/blocks/info-block/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./info-block"
