+++ src/components/blocks/info-block/info-block.module.css
@@ -17,7 +17,7 @@
   width: 170px;
 }
 
-.participation {
+.list {
   display: flex;
   width: 50%;
 }
