+++ src/pages/project-page/project-page.tsx
@@ -4,7 +4,7 @@ import Crumbs from "../../components/crumbs"
 import { MOCK_PROJECTS } from "../../mock/mock-projects"
 import { createCrumps } from "./helpers"
 import FilterIconBlock from "../../components/filter-icon-block"
-import ProjectInfo from "../../components/project-info"
+import ProjectInfo from "../../components/blocks/project-info"
 import { BLOCK_TYPE } from "../../shared/constants"
 import Carousel from "../../components/blocks/carousel"
 import TextColumns from "../../components/blocks/text-columns"
